#1384 Fixes towards the review
