+++ src/psyclone/tests/psyir/nodes/omp_directives_test.py
@@ -594,10 +594,11 @@ def test_omp_taskloop_node_str():
     assert expected_output in out
 
 
-@pytest.mark.parametrize("grainsize,num_tasks,nogroup", [(None, None, False),
-                                                         (32, None, True),
-                                                         (None, 32, True)])
-def test_omp_taskloop_gencode(grainsize, num_tasks, nogroup):
+@pytest.mark.parametrize("grainsize,num_tasks,nogroup,clauses",
+                         [(None, None, False, ""),
+                          (32, None, False, " grainsize(32)"),
+                          (None, 32, True, " num_tasks(32), nogroup")])
+def test_omp_taskloop_gencode(grainsize, num_tasks, nogroup, clauses):
     '''Check that the gen_code method in the OMPTaskloopDirective
     class generates the expected code. Use the gocean API.
     '''
@@ -618,55 +619,6 @@ def test_omp_taskloop_gencode(grainsize, num_tasks, nogroup):
     goceantrans.apply(schedule.children[0])
 
     code = str(psy.gen)
-
-    clauses = ""
-    if grainsize is not None:
-        clauses = " grainsize({0})".format(grainsize)
-        if nogroup:
-            clauses = clauses + ", nogroup"
-    if num_tasks is not None:
-        clauses = " num_tasks({0})".format(num_tasks)
-        if nogroup:
-            clauses = clauses + ", nogroup"
-
-    assert (
-        "    !$omp parallel default(shared), private(i,j)\n" +
-        "      !$omp master\n" +
-        "      !$omp taskloop{0}\n".format(clauses) +
-        "      DO" in code)
-    assert (
-        "      END DO\n" +
-        "      !$omp end taskloop\n" +
-        "      !$omp end master\n" +
-        "      !$omp end parallel" in code)
-
-    assert taskloop_node.begin_string() == "omp taskloop{0}".format(clauses)
-
-
-def test_omp_taskloop_apply_gencode():
-    '''Check that the gen_code method in the OMPTaskloopDirective
-    class generates the expected code when passing options to
-    the OMPTaskloopTransformation apply. Use the gocean API.
-    '''
-    _, invoke_info = parse(os.path.join(GOCEAN_BASE_PATH, "single_invoke.f90"),
-                           api="gocean1.0")
-    taskloop = OMPTaskloopTrans()
-    master = OMPMasterTrans()
-    parallel = OMPParallelTrans()
-    psy = PSyFactory("gocean1.0", distributed_memory=False).\
-        create(invoke_info)
-    schedule = psy.invokes.invoke_list[0].schedule
-
-    taskloop.apply(schedule.children[0], {"nogroup": True})
-    taskloop_node = schedule.children[0]
-    master.apply(schedule.children[0])
-    parallel.apply(schedule.children[0])
-    goceantrans = GOceanExtractTrans()
-    goceantrans.apply(schedule.children[0])
-
-    code = str(psy.gen)
-
-    clauses = " nogroup"
 
     assert (
         "    !$omp parallel default(shared), private(i,j)\n" +
